adjust FlfGameObj + Game splits

+++ src/nw4r/math/math_constant.h
@@ -10,7 +10,7 @@
  *
  ******************************************************************************/
 // Pi mathematical constant
-#define NW4R_MATH_PI M_PI
+#define NW4R_MATH_PI (float)M_PI
 
 // ln(2)
 #define NW4R_MATH_LN_2 0.69314718056f
